aead: Rename StatelessAead -> Aead; Aead -> AeadMut (closes #46)
As it seems like most commonly used AEAD algorithms can be implemented
as what was previously called `StatelessAead`, this commit renames it to
give it the shorter, more convenient name (`Aead`).

Additionally, it renames the prior `Aead` trait to `AeadMut`, as
suggested by @burdges in #46.

This seems to follow the general Rust convention of making the immutable
version of something shorter and easier to use, and explicitly calling
out `Mut` where it exists.