aead: Add in-place API based on `Buffer` trait
Defines a `Buffer` trait for `Vec`-like types which impl `AsRef<[u8]>`,
`AsMut<[u8]>`, `extend_from_slice()`, and `truncate()`, and optionally
impls it for `Vec<u8>` (when the `alloc` feature is enabled), along with
the option to use `heapless::Vec` (when the `heapless` feature is
enabled), providing a true `#![no_std]`-friendly option.

Uses `impl Buffer` as the argument for `encrypt_in_place` and
`decrypt_in_place`, with default implementations that assume a postfix
authentication tag. This allows these methods to handle ciphertext
assembly and parsing, after which they can invoke the `_detached` APIs.