signature: design documentation
This commit adds high-level descriptions of how the `signature` crate is
designed, including alternatives considered, and hopefully details why
it is the way it is.

These notes also include a rough sketch of potential additional future
traits which could better optimize for implementers of signature traits,
rather than consumers (with a blanket impl to link the two).