signature: remove BoxError type alias
We can revisit adding this back after 1.0, but it seems possible such a
type may eventually be stabilized in `std`.

It's better to "wait and see" if that happens and add `BoxError` back in
the event there's a desire for more ergonomic boxed errors for `std`
users rather than prematurely stabilizing it if it turns out to be
unnecessary.