Uprev generic array to version 0.14
In 0.14 they added a pretty nice feature, which is that you can
now split not just arrays that you hold by value, but you can
split references to generic arrays, into a pair of references to
generic arrays corresponding to the subslices, where the size
calculations are done correctly at compile-time and the buffer
size information is all preserved.

https://docs.rs/generic-array/0.14.1/src/generic_array/sequence.rs.html#302-320

I have some ECIES code that looks like this, that builds on `aead` trait:

```
    fn decrypt_in_place_detached(
        &self,
        key: &RistrettoPrivate,
        footer: &GenericArray<u8, Self::FooterSize>,
        buffer: &mut [u8],
    ) -> Result<(), Error> {
        let (footer, version_data): (_, &GenericArray<u8, U2>) = footer.split();
```

Alternate version is like

```
    fn decrypt_in_place_detached(
        &self,
        key: &RistrettoPrivate,
        footer: &GenericArray<u8, Self::FooterSize>,
        buffer: &mut [u8],
    ) -> Result<(), Error> {
        let (footer, version_data) = Split::<u8, U48>::split(footer);
```

I think these are both expected to work generic array 0.14. There
are alternatives of course but they are less tidy :)

Since I'm getting the generic array pub export from you, I think
that to use this, I have to convince you that it's worth it to uprev.
LMK what you think! Thanks