Add "aead::encrypt_fixed_length" and "aead::decrypt_fixed_length"
I wanted to suggest adding these two APIs which provide alternative
ways to use aead in a no-alloc build environment, without using the
heapless library. Instead it uses `GenericArray::{split, concat}`
to put the tag at the end of the buffer.

I liked using this API in the context when, a struct containing
fixed length buffers is where we need to place the ciphertext.

The mc-crypto-box trait is based closely on this aead trait, and
that's where I got experience using this version of the `encrypt`
function:

https://github.com/mobilecoinofficial/mobilecoin/blob/master/crypto/box/src/traits.rs#L140

I thought I would propose this extension upstream here.