cryptography: fix rustdoc by linking directly to https://docs.rs (#202)
Unfortunately even after #200 the links to the documentation for the
various trait crates is still broken in the exact same way. I suspect
this might be some kind of rustdoc/docs.rs bug around re-exports.

This commit explicitly links to https://docs.rs for each of the crates
rather than using *any* kind of indirection, in hopes that corrects the
issue.