elliptic-curve: replace FixedBaseScalarMul with ops::MulBase
`FixedBaseScalarMul` was a hack around the lack of an associated
`Scalar` type on a curve and generic way of converting a `SecretKey` to
a `Scalar`.

We have both of those things, so this commit replaces it with
an `ops::MulBase` trait intended to be impl'd on `Scalar` types.

Further, it bounds the `Arithmetic::Scalar` associated type on this,
which allows it to be used for generic `SecretKey` -> `PublicKey`
conversions.

It also means the `Arithmetic` trait now has a(n implicit) way of
representing the generator point, making it genuinely useful for generic
arithmetic operations. This would also be useful for e.g. a generic
implementation of Diffie-Hellman.