elliptic-curve: basic OID support
Adds a (very simple, const-and-no_std-friendly) `ObjectIdentifier` type
and an `Identifier` trait for attaching an OID to a curve.

No validation of the nodes is presently performed (and since it's impl'd
as a `const fn`, it'd require a MSRV bump to do so). Also requires the
OID be specified as an array literal: no parsing is supported for either
string or binary-encoded OIDs.

Only string serialization is supported for now. A `no_std`-friendly
binary serializer will be somewhat tricky.