elliptic-curve: NonZeroScalar + Generator (replaces MulBase)
Adds a `NonZeroScalar` type generic over curves along with a `Generator`
trait which together with the `Mul` trait from `core` can replace the
`MulBase` trait.

`NonZeroScalar` is able to use the bounds on `Arithmetic::Scalar` and
therefore usable anywhere a `Scalar` otherwise would, but with the
guarantee that multiplication by an `AffinePoint` will not result in the
point at infinity, for which no `AffinePoint` representation is possible
given current trait bounds/implementations.