[WIP] elliptic-curve: refactor `PublicKey` into `sec1::EncodedPoint`
Previously the `PublicKey` was a "wire type" for SEC1 encoded curve
points which by design didn't validate the point is on the curve (so as
to avoid mandating a curve arithmetic implementation for the type to be
useful).

However, this has a pretty big drawback, in that when we do have a curve
arithmetic implementation, there are many fallible operations which
could be infallible if we knew the point is valid a priori.

Additionally, thanks to `typenum`-based arithmetic, it has pretty
atrocious trait bounds, which are needed anywhere we presently deal with
a public key.

This commit moves this encoded point type to `sec1::EncodedPoint`, which
both much more precisely describes what it represents, but also moves
it out-of-the-way so we can add a better `PublicKey` type (e.g. one
which is or wraps an `AffinePoint` in crates which have a curve
arithmetic implementation).