elliptic-curve: refactor `PublicKey` into `sec1::EncodedPoint` (#264)
elliptic-curve: refactor `PublicKey` into `sec1::EncodedPoint`

Moves the type for SEC1 encoded points to `sec1::EncodedPoint`, which
both much more precisely describes what it represents, but also moves it
out-of-the-way so we can add a better `PublicKey` type (e.g. one which
is or wraps an `AffinePoint` in crates which have a curve arithmetic
implementation).

It also changes `sec1::EncodedPoint` to be backed by a byte slice which is
always the size of a SEC1 encoded uncompressed point, eliminating the
previous `CompressedPoint` and `UncompressedPoint` types. This type
can still contain a SEC1-encoded compressed point, sliced to the proper
size via the `as_bytes()` method.

This allows for simplified trait bounds as we no longer need to care
about `CompressedPointSize` in the bounds.