elliptic-curve: import arithmetic helper functions
These are helper functions for performing arithmetic on field element
"limbs", e.g. implementing carry chains.

The particular versions in this file come from the `k256`, which started
as a fork of the `p256` crate, and notably provide 32-bit and 64-bit
versions of the same functions.

These are potentially helpful to have in the `elliptic-curve` crate for
several reasons, most notably for testing things which are generic
across elliptic curves without having to provide a full curve arithmetic
implementation.

Also having things in one place is nice, and potentially we can use
these to provide a generic baseline implementation of certain types of
curve arithmetic for curves that don't provide their own arithmetic.