elliptic-curve: add `ff` and `group` crate dependencies
Adds the `ff` and `group` crates as dependencies, with the goal of
replacing some of the existing traits and trait relationships in
`elliptic-curve` with ones from `ff`/`group`.

This PR also removes the `rand` and `weierstrass` features from the
`elliptic-curve` crate, making `rand_core` a hard requirement, since it
seems at least `rand_core` will probably be a hard requirement for
`group`, and requiring an RNG as a hard dependency makes sense.