elliptic-curve: vendor core `group` traits; add `ff` dependency
Vendors the core set of group traits from the `group` crate:

https://github.com/zkcrypto/group

Using the upstream crate is presently blocked on `no_std` support:

https://github.com/zkcrypto/group/pull/9

However, aside from a suggested change to the RNG API
(zkcrypto/group#10) the traits used in this PR are verbatim from the
upstream crate, so migrating to it should be easy.

This PR also removes the `rand` and `weierstrass` features from the
`elliptic-curve` crate, making `rand_core` a hard requirement, since it
seems at least `rand_core` will probably be a hard requirement for
`group`, and requiring an RNG as a hard dependency makes sense.