elliptic-curve: add `arithmetic` feature
This crate now has a large(r) number of dependencies via the `ff` and
`group` crates (notably the recently introduced `bitvec` and its
transitive dependencies).

This commit adds an `arithmetic` feature and gates related functionality
under it, along with the `ff` and `group` crate dependencies.

This ensures these dependencies (and their transitive dependencies) can
be excluded when the `arithmetic` features of elliptic curve
implementations are non-existent or an associated feature disabled.