elliptic-curve: refactor ProjectiveArithmetic trait
Renames the `Arithmetic` trait to `ProjectiveArithmetic` and changes it
to have only a single associated type: `ProjectivePoint`.

Adds `AffinePoint<C>` and `Scalar<C>` type aliases which are able to
look up the corresponding associated type chain (along with
`ProjectivePoint<C>` for consistency, although this is available as
`<C as ProjectiveArithmetic>::ProjectivePoint`).