cryptography: add `std` feature
Adds an off-by-default `std` feature which activates the `std` features
of all of the other dependencies.

This is a somewhat lousy way of doing this, and I wish that there were a
better approach, so much that I posted a Pre-Pre-RFC about it:

https://internals.rust-lang.org/t/pre-pre-rfc-weak-cargo-feature-activation/13141

For now though, this is immediately useful to me.