elliptic-curve: add SecretBytes newtype
The `SecretValue` trait was supposed to allow curve crates with an
`arithmetic` feature to use a `NonZeroScalar` as their secret value,
and crates which don't to use `FieldBytes`.

Unfortunately `zeroize` has no `generic-array` integration, and
therefore `GenericArray` (which is what `FieldBytes` is a type alias
for) cannot directly be used as `SecretValue::Secret`.

This commit adds a thin `SecretBytes` newtype which wraps `FieldBytes`
and impls `Zeroize` so it can be used as `SecretValue::Secret`.