stream-cipher: remove NewBlockCipher bound on FromBlockCipher
Note: breaking change.

As far as I can tell this was an oversight: `FromBlockCipher` never
needs to instantiate a new block cipher (isn't that the point?),
making this bound needlessly limiting.

The rationale for it in the first place appears to be for the blanket
impl of `NewStreamCipher` for `C: FromBlockCipher` which needs it, but
we can add the bound to the blanket impl, rather than (needlessly)
sticking it on the trait.