stream-cipher: add FromBlockCipherMut trait (#334)
Support for initializing a stream cipher from a `BlockCipherMut`.

This is a corollary to `FromBlockCipher`, but for cases where `mut`
access is needed to the inner block cipher instance (e.g. cryptographic
accelerator hardware peripherals).

Using chained blanket impls, it's possible to allow a blanket impl of
`NewStreamCipher` for types which impl either `FromBlockCipherMut` or
`FromBlockCipher`.