cipher: re-export `generic_array` from toplevel (#343)
...along with `typenum::consts`

Failure to do this was an oversight: it doesn't really make sense to
re-export these twice from both the `block` and `stream` modules.

That said, `cipher` v0.2 is already shipped, and it does ease a
sed-based migration to keep these re-exports for now.

This commit adds a third re-export from the toplevel, and comments to
remove the re-exports from the `block` and `stream` crates. This should
happen before `cipher` v0.3 is released.