cipher: revamp macro names
The old macro names come from the `block-cipher` and `stream-cipher`
crates, where the crate names disambiguated them.

However, in the same crate, and re-exported from the toplevel as macros
are, it's now unclear whether a macro is for a block cipher or a stream
cipher.

This commit adds `block_cipher*` and `stream_cipher` to the start of
each macro name to disambiguate them.

It also maintains the existing macros as wrappers for the new names, but
with a `#[deprecated]` attribute added.