Bump `const-oid` to v0.2; MSRV 1.46+
As MSRV bump is considered a breaking change; this also bumps the
Cargo.toml version to v0.7.0-pre.

This version of `const-oid` uses the new `const fn` features to provide
some basic OID validation.

While this alone may be a silly reason to bump MSRV to 1.46+, many of
the elliptic curve crates could also benefit from using `const fn` in
a similar manner, so bumping MSRV is generally useful.