elliptic-curve: remove `subtle` usage on sec1::EncodedPoint
From a practical perspective, `sec1::EncodedPoint` is always used with
public values.

The previous use of `CtOption` was almost exclusively for error
handling, and in every case immediately wrapped in an `Option` or
`Result` to avoid end users dealing with public keys from having to deal
with `subtle`.

Removing the usage of `subtle` in these cases dramatically simplifies
the code and required trait bounds.