elliptic-curve: remove Deref impl from PublicKey; add ::as_affine()
Types with `Deref` impls shouldn't have inherent methods, however
`PublicKey` has several.

This makes `PublicKey` more than a "smart pointer" type which `Deref` is
intended to be used with: it handles encoding and conversions in a way
that can be shared across curves.

It already provided an `AsRef` impl for obtaining the `AffinePoint`,
however this PR also adds a new inherent `as_affine` method as well.