elliptic-curve: make `ToEncodedPoint` fallible
Closes #399.

As discussed in the aforementioned issue, this changes `ToEncodedPoint`
to be fallible, returning an `Option<EncodedPoint<C>>` with the
expectation that `None` is returned for the additive identity
(a.k.a. point-at-infinity)

Brief background: originally it was not expected for `AffinePoint` to be
able to encode the additive identity, however supporting it allowed for
infallible conversions between `AffinePoint` <-> `ProjectivePoint`.

Technically, according to SEC1 Section 2.3.3, the identity can be
represented by the `Elliptic-Curve-Point-to-Octet-String` encoding by
using a single NULL byte (i.e. `[0u8]`).

However, this would have the disadvantage of complicating
`EncodedPoint`, making APIs that are presently infallible fallible, and
may be unexpected by end users of the traits.

If there's a legitimate use case for supporting SEC1-encoded identity
points, we can revisit this tradeoff, adding support for encoding the
identity element to `EncodedPoint`, and making `ToEncodedPoint`
infallible once again.