elliptic-curve: add support for SEC1 identity encoding
Closes #399.

The `ToEncodedPoint` trait is infallible. We can either make it fallible
(as attempted in #400) or add support for the SEC1 encoding for the
identity point.

This commit adds SEC1 support for identity, allowing `ToEncodedPoint`
to remain infallible.

Unfortunately breaking as it requires adding new enum variants and some
method signature changes.

However, that said, this should make the SEC1 implementation "complete"
in that it implements the full `Elliptic-Curve-Point-to-Octet-String`
encoding (and decoding).