elliptic-curve: `dev` module
There is presently no good way to test traits and types in this crate
which depend on a concrete curve implementation.

This commit adds a `dev` module gated on an eponymous crate feature
which provides a `MockCurve` type. This type does not provide a full
curve arithmetic implementation (most methods are stubbed out using
`unimplemented!`), but rather implements the full set of traits used in
the bounds of traits defined in the `elliptic-curve` crate.

This type is actually an extraction from the `ecdsa` trait, where
similar testing problems were previously encountered. However, now that
there's a `dev` module, the `ecdsa` crate can now use it!