elliptic-curve: regression test for PublicKey::from_encoded_point
The invariant of `PublicKey` is that the inner affine point is *NOT*
the identity, however there was a bug in the `FromEncodedPoint`
implementation which was fixed in #416.

However, at the time we couldn't add a regression test because there was
no concrete curve type to use for writing such a test.

A `MockCurve` was added in #419, so this commit uses it to write such
a regression test.