cipher: revert SeekNum::from_block_byte change
This reverts the implementation of `SeekNum::from_block_byte` which was
merged as part of #435.

I'm not exactly sure what the issue is and it's somewhat difficult to
debug in that it's code that involves both generics and macros causing
an error as part of any failure in a long chain of checked arithmetic,
which I'm trying to debug from the context of a concrete stream cipher
impl (both `chacha20` and `salsa20`) where it's operating over a generic
type.

The error manifests as `OverflowError`:

https://github.com/RustCrypto/stream-ciphers/pull/205#issuecomment-752314229

This commit reverts to the previous implementation, which is at least
much simpler.