aead: `stream` module (#436)
Implementation of the STREAM construction as described in the paper
"Online Authenticated-Encryption and its Nonce-Reuse Misuse-Resistance":

https://eprint.iacr.org/2015/189.pdf

The implementation is generic over AEAD ciphers and is factored into
a low-level `StreamPrimitive` trait (permitting different "flavors"
of STREAM) as well as higher-level stateful `Encryptor` and `Decryptor`
objects which are generic over `StreamPrimitive` types.

Includes two concrete implementations of `StreamPrimitive`:

- `StreamBE32`: the original version of STREAM described in the paper,
  with a nonce in the form: `prefix || counter || last_block`, where
  `counter` is a 32-bit big endian-encoded integer, and `last_block`
  is a 1-byte flag.
- `StreamLE31`: uses a 31-bit counter and 1-bit last block flag,
  packed into the last 4 bytes of the nonce as a little endian integer.

Using little endian provides better performance on commonly used CPU
architectures, and using a 1-bit last block flag ensures the user-facing
STREAM nonce is even numbered in terms of bytes (e.g. for a 96-bit nonce
it'd be 64-bits or 8-bytes instead of a 7-byte nonce using the
construction described in the paper) and also avoids wasting bits.