password-hash: remove B64 restrictions on Salt
Changes the `Salt` type to be a simple newtype of `ValueStr`.

Since the `Salt` value is now borrowed, this unfortunately removes the
ability to generate a random salt into a buffer, so this PR also removes
the `rand_core` dependency.

We can potentially revisit adding a type like `SaltBuf` to use for an
owned salt value.