password-hash: ParamsString refactor
Replaces a structured type with a builder for params strings which can
also parse the in-place via an iterator.

The previous implementation had a size of 320(!) bytes despite the fact
all of the string data was borrowed. It's considerably more efficient to
have a buffer type that can build the parameter string which is
internally stored as a byte array.

The new implementation uses a 127-byte array with a 1-byte length for a
struct which is a total of 128-bytes.

This length was picked in part by adding headroom to a maximally long
Argon2 params string, which looks roughly like this:

    m=4294967296,t=4294967296,keyid=XXXXXXXXXXX,data=XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX

This string contains all Argon2 parameters at maximum length which comes
to a total of 92-bytes, which means 127 provides 35 bytes of headroom.

The implementation is written in such a way that it should be easy to
swap out a `String` for backing storage when the `alloc` feature is
enabled, if it turns out the 128-byte size is suboptimal.