elliptic-curve: remove FromDigest trait
As discussed in #481, this removes the `FromDigest` trait, and with it
any dependency on the `digest` crate (at least for now).

The reasoning is `FromDigest` in its current form exists to solve a
single concern: hash-to-scalar for the specific purposes of ECDSA, which
is implemented in the `k256` and `p256` crates as a "narrow" reduction
per the requirements of ECDSA.

Given that, this trait is best moved to the `ecdsa` crate, clearing the
way for a move generally useful, well-considered, and well-specified and
standardized hash-to-curve construction as requested in #481.

The plan after this is to provide an identical trait, but in the `ecdsa`
crate instead of the `elliptic-curve` crate.