elliptic-curve: fix `SecretKey::from_bytes` slice parsing
Closes #588

`SecretKey::from_bytes` was previously attempting to use
`TryFrom<&[u8]>` on `GenericArray`, however `GenericArray` actually has
a panicking `From` impl, and was therefore receiving panicking `TryFrom`
via the blanket impl of `TryFrom` for `From`.

This commit adds a regression test which captures the bug, and fixes the
implementation to test that the slice is correctly sized.