elliptic-curve: add Order trait
Adds a trait which notably does *not* depend on the `arithmetic` feature
which allows associating a constant for the curve's order with a
particular curve type.

Since it stores the curve type as a constant, it's not possible to use
`GenericArray`, so instead it has each curve specify a `Limbs` type
which is used as the type for an `ORDER` constant.

The longer-term goal will be to merge this with the `Curve` trait,
ensuring that every curve implementation has a known order regardless of
whether it provides an arithmetic backend or not.

This can be used as the foundation for making certain functionality
generic across curve implementations at a baseline, such as checking if
scalars are in-range, and as such can simplify things which are
presently conditionally defined based on the presence or absence of an
arithmetic backend.