password-hash: use `subtle` crate for comparing hash `Output`
Previously a non-short-circuiting comparison was used, however the
`subtle` crate provides a more robust option in this regard with the
`ConstantTimeEq` trait and its associated impls on byte slices.

This commit also includes an updated rationale for why password hashes
benefit from constant time comparisons, including a description of a
timing attack on password hash verification.