elliptic-curve: merge `Curve` and `Order` traits
The `Order` trait was additively retrofitted in a non-breaking way.
However all of the https://github.com/rustcrypto/elliptic-curves crates
now impl it, and thus it's ready to merge into the `Curve` trait proper
as a mandatory-to-impl feature.

The current form is also a stopgap: now that the `crypto-bigint` crate
has an initial release, it can be used as the type to store the
`Curve::ORDER` in a const-friendly way, and also it can leverage the
`crypto-bigint` crate's `generic-array` interop to replace the
`FieldSize` constant as well.

This is the first step towards those refactorings.