elliptic-curve: use `crypto-bigint` to represent `Curve::Order` (#645)
The `crypto_bigint::UInt` type has const-friendly initializers that can
parse the curve order from e.g. hexadecimal.

This commit changes the `Order` to be represented as a `UInt` type.
This also permits things like simplified (and constant time, if needed)
comparisons to ensure a given value is in range.

Additionally, this commit reimplements the `dev` module's `Scalar` type
for `MockCurve` to use a `crypto_bigint::U256` internally.