elliptic-curve: make ScalarBytes<C> the SecretKey<C> internal repr
Previously SecretKey<C> had a generic internal value, depending on
whether or not the `arithmetic` feature was activated. This was mostly
trying to work around the fact that there wasn't a baseline arithmetic
impl available to validate secret keys.

Now there is! SecretBytes<C> uses C::UInt and its may features as a
`crypto-bigint` to always validate the internal bytes.

This means we can simply use `SecretBytes<C>` as the inner value of a
`SecretKey<C>`, and still provide infallible conversions to
`NonZeroScalar<C>`.