elliptic-curve: add TODO for replacing UInt with an associated constant
Ideally instead of specifying a `Curve::UInt` associated type, we could
instead specify a constant, and then pass that as a const generic
parameter to `crypto_bigint::UInt`.

Unfortunately that requires a lot more functionality than is implemented
in `min_const_generics`:

https://github.com/rust-lang/rust/issues/60551

This commit makes a note of as much in the comments, calling it out as
future work.