elliptic-curve: bump `crypto-bigint` dependency to v0.2.4
This release includes some new features which can be used to calculate
various constants from an elliptic curve's order using `const fn`s.

Bumping the minimum release allows for elliptic curve libs to pin to a
release of the `elliptic-curve` crate which includes this new version,
allowing them to ensure the transitive dependency is up-to-date without
explicitly pinning to it.