elliptic-curve: add `ScalarCore<C>` type
This is incremental work towards having a "one true scalar" type which
is generic over elliptic curves, using `crypto-bigint` for the backing
arithmetic.

It's presently called `ScalarCore<C>` to prevent it from conflicting
with the existing `Scalar<C>`, however the goal is to completely replace
the current `Scalar<C>`.

This commit also does the following:

- Removes the `ScalarBytes<C>` type as `ScalarCore<C>` can subsume its
  previous use cases
- Makes the `zeroize` crate a hard dependency (i.e. removes the
  `zeroize` feature, making it mandatory)