elliptic-curve: use `sec1` crate's `EncodedPoint`
The `EncodedPoint` type was extracted to the `sec1` crate in:

https://github.com/RustCrypto/formats/pull/45

This switches the `elliptic-curve` to using the newly extracted
implementation.

This was a bit tricky/invasive as now `EncodedPoint` now only knows a
size and not the rest of the details of the curve.

APIs that were previously heavily leveraging `EncodedPoint` should
probably switch to the `PublicKey` type instead unless they are
specifically intending to work with SEC1-encoded points.