elliptic-curve: remove narrow `Reduce` bound on `Scalar`
Narrow reductions are not ideal and introduce much more bias than a wide
reduction. If we are going to bound on a `Reduce` impl, it should at
least include a wide reduction so we don't accidentally drive people to
narrow reductions instead.

While they are needed for ECDSA, narrow reductions should probably be
discouraged in other non-legacy applications.