elliptic-curve: bump MSRV to 1.56
The RustCrypto/formats repo was bumped to Rust 2021 edition, and some of
those crates (e.g. `der`, `sec1`) are hard dependencies for the
`elliptic-curve` crate:

https://github.com/RustCrypto/formats/pull/136

This is a corresponding Rust 2021 edition bump for the `elliptic-curve`
crate, as well as the omnibus `crypto` crate whose MSRV is determined by
the highest MSRV crate in this repo.