password-hash: better `Debug`/`Display` impls for `SaltString`
Adds impls of `Debug` and `Display` on `SaltString` similar to the ones
on `Salt` (the reference type).

Previously there was no `Display` impl at all.

The `Debug` impl was previously derived, and therefore had confusing
output due to the use of a fixed-sized internal buffer to support
heapless platforms.

Now they both print the string value, with the `Debug` impl wrapped in
`SaltString(...)`.