elliptic-curve: add `AffineXCoordinate` trait
Protocols like ECDH and ECDSA rely on x-coordinate access.

We've tried various ways of trying to keep affine coordinate access out
of the public API, but it complicates generic implementations of these
protocols.

Most recently RustCrypto/signatures#395 attempted to add a trait which
encapsulated ECDSA signing, but a similar (or expaned version of the
same) trait is needed for verification.

All of these problems go away, and can be expressed with simple trait
bounds, if there is some way to access the x-coordinate of an
`AffinePoint` as serialized `FieldBytes`.

This commit adds such a trait: `AffineXCoordinate`, which is a mandatory
bound of `AffineArithmetic::AffinePoint`.