elliptic-curve: `non_zero` => `nonzero`
In #827, a `ReduceNonZero` trait was added whose method name uses an
underscored `*_non_zero`.

However, there's already one bit of precedent in the public API:

- `SecretKey::to_nonzero_scalar`

Since this is already a stable API and providing precedent, for
consistency this commit changes all of the unstable/unreleased APIs and
internal module naming to use `nonzero` instead of `non_zero`, as this
is a non-breaking change.