Improve API for hash2curve functions (#876)
- Allow methods to accept multiple `data`/`msg`s by using `&[&[u8]]`, á la `Hkdf::expand_multi_info()`
- Implement `hash_to_scalar`, got an implementation for P-256 already working with test vectors.
- Remove `'static` requirement for `dst`.

The goal here is to make this usable for voprf. VOPRF sometimes builds complex `msg`s that are impossible to concat without allocation, this is why the multi-message API was necessary. I would even prefer just to take an `Iterator<Item = &[u8]>`.

The `'static` requirement doesn't work for VOPRF because it needs to build the `dst` from multiple parts that are party derived from associated types from traits, this can't be used in const to build a `&'static [u8]` in Rust right now.