elliptic-curve: v0.11.11 changelog entry
It looks like the docs.rs build for v0.11.10 failed due to nightly bugs:

https://docs.rs/crate/elliptic-curve/0.11.10/builds/500473

This adds a changelog entry for a v0.11.11 release, which is unchanged
from v0.11.10, but released again to trigger a new docs.rs build.

The `master` branch is already the v0.12 prerelease series, so this
commit doesn't contain a full release.